requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.2.2
channels_redis==4.3.0
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==5.1.3
django-cors-headers==4.6.0
django-redis==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
msgpack==1.1.1
mysqlclient==2.2.6
packaging==24.2
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.0
pyOpenSSL==25.1.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==6.2.0
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.2
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2024.2
uritemplate==4.1.1
zope.interface==7.2
